Additional changes and files for print function
