+++ Team-C/test_wizard/lib/widgets/edit_prompt.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class EditPrompt extends StatelessWidget {
+  const EditPrompt({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(top: 20),
+      child: Center(
+        child: ElevatedButton(
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Colors.white,
+            foregroundColor: const Color(0xff0072bb),
+          ),
+          onPressed: () {},
+          child: const Text('Edit Prompt (Advanced)'),
+        ),
+      ),
+    );
+  }
+}
